Rename directory and file
Renamed from general to particular, so that the user can find its way to specific issue. For
example, installing-anaconda-etc is now anaconda-install-etc, so that the alphabetical order
organizes all issues with anaconda (for example, also adding it to the path).
This should prove convenient when we gather issues regarding other software than anaconda.